fix(client): hoist fetch
