+++ client/cmd/demarkus-mcp/main.go
@@ -13,14 +13,14 @@ import (
 	"strings"
 	"time"
 
+	"github.com/latebit/demarkus/client/fetch"
 	"github.com/latebit/demarkus/client/internal/cache"
-	"github.com/latebit/demarkus/client/internal/fetch"
 	"github.com/latebit/demarkus/client/internal/graph"
 	"github.com/latebit/demarkus/client/internal/graphstore"
 	"github.com/latebit/demarkus/client/internal/index"
 	"github.com/latebit/demarkus/client/internal/links"
-	"github.com/latebit/demarkus/client/internal/merge"
 	"github.com/latebit/demarkus/client/internal/tokens"
+	"github.com/latebit/demarkus/client/merge"
 	"github.com/latebit/demarkus/protocol"
 	"github.com/mark3labs/mcp-go/mcp"
 	mcpserver "github.com/mark3labs/mcp-go/server"
