+++ client/cmd/demarkus-mcp/main_test.go
@@ -9,7 +9,7 @@ import (
 
 	"path/filepath"
 
-	"github.com/latebit/demarkus/client/internal/fetch"
+	"github.com/latebit/demarkus/client/fetch"
 	"github.com/latebit/demarkus/client/internal/graph"
 	"github.com/latebit/demarkus/client/internal/graphstore"
 	"github.com/latebit/demarkus/protocol"
