+++ client/internal/fedcrawl/crawl.go
@@ -11,7 +11,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/latebit/demarkus/client/internal/fetch"
+	"github.com/latebit/demarkus/client/fetch"
 	"github.com/latebit/demarkus/client/internal/index"
 	"github.com/latebit/demarkus/client/internal/links"
 	"github.com/latebit/demarkus/client/internal/tokens"
