+++ client/internal/graphstore/store.go
@@ -18,7 +18,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/latebit/demarkus/client/internal/fetch"
+	"github.com/latebit/demarkus/client/fetch"
 	"github.com/latebit/demarkus/client/internal/graph"
 	"github.com/latebit/demarkus/client/internal/tokens"
 )
